Update reference assemblies
