+++ src/Components/Authorization/ref/Microsoft.AspNetCore.Components.Authorization.csproj
@@ -4,15 +4,13 @@
     <TargetFrameworks>netstandard2.0;netcoreapp3.0</TargetFrameworks>
   </PropertyGroup>
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <Compile Include="Microsoft.AspNetCore.Components.Web.netstandard2.0.cs" />
+    <Compile Include="Microsoft.AspNetCore.Components.Authorization.netstandard2.0.cs" />
+    <Reference Include="Microsoft.AspNetCore.Authorization"  />
     <Reference Include="Microsoft.AspNetCore.Components"  />
-    <Reference Include="Microsoft.AspNetCore.Components.Forms"  />
-    <Reference Include="Microsoft.Extensions.DependencyInjection"  />
   </ItemGroup>
 <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.0'">
-    <Compile Include="Microsoft.AspNetCore.Components.Web.netcoreapp3.0.cs" />
+    <Compile Include="Microsoft.AspNetCore.Components.Authorization.netcoreapp3.0.cs" />
+    <Reference Include="Microsoft.AspNetCore.Authorization"  />
     <Reference Include="Microsoft.AspNetCore.Components"  />
-    <Reference Include="Microsoft.AspNetCore.Components.Forms"  />
-    <Reference Include="Microsoft.Extensions.DependencyInjection"  />
   </ItemGroup>
 </Project>
