+++ src/Components/Blazor/Blazor/ref/Microsoft.AspNetCore.Blazor.csproj
@@ -7,5 +7,6 @@
     <Compile Include="Microsoft.AspNetCore.Blazor.netstandard2.0.cs" />
     <Reference Include="Mono.WebAssembly.Interop"  />
     <Reference Include="Microsoft.AspNetCore.Components.Web"  />
+    <Reference Include="Microsoft.Extensions.Options"  />
   </ItemGroup>
 </Project>
