+++ src/Components/Forms/ref/Microsoft.AspNetCore.Components.Forms.csproj
@@ -4,15 +4,12 @@
     <TargetFrameworks>netstandard2.0;netcoreapp3.0</TargetFrameworks>
   </PropertyGroup>
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <Compile Include="Microsoft.AspNetCore.Components.Web.netstandard2.0.cs" />
+    <Compile Include="Microsoft.AspNetCore.Components.Forms.netstandard2.0.cs" />
     <Reference Include="Microsoft.AspNetCore.Components"  />
-    <Reference Include="Microsoft.AspNetCore.Components.Forms"  />
-    <Reference Include="Microsoft.Extensions.DependencyInjection"  />
+    <Reference Include="System.ComponentModel.Annotations"  />
   </ItemGroup>
 <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.0'">
-    <Compile Include="Microsoft.AspNetCore.Components.Web.netcoreapp3.0.cs" />
+    <Compile Include="Microsoft.AspNetCore.Components.Forms.netcoreapp3.0.cs" />
     <Reference Include="Microsoft.AspNetCore.Components"  />
-    <Reference Include="Microsoft.AspNetCore.Components.Forms"  />
-    <Reference Include="Microsoft.Extensions.DependencyInjection"  />
   </ItemGroup>
 </Project>
