+++ src/Components/Server/ref/Microsoft.AspNetCore.Components.Server.csproj
@@ -5,6 +5,7 @@
   </PropertyGroup>
   <ItemGroup Condition="'$(TargetFramework)' == 'netcoreapp3.0'">
     <Compile Include="Microsoft.AspNetCore.Components.Server.netcoreapp3.0.cs" />
+    <Reference Include="Microsoft.AspNetCore.Components.Authorization"  />
     <Reference Include="Microsoft.AspNetCore.Components.Web"  />
     <Reference Include="Microsoft.AspNetCore.DataProtection"  />
     <Reference Include="Microsoft.AspNetCore.SignalR"  />
