+++ src/Mvc/Mvc.ViewFeatures/ref/Microsoft.AspNetCore.Mvc.ViewFeatures.csproj
@@ -12,5 +12,6 @@
     <Reference Include="Microsoft.AspNetCore.Html.Abstractions"  />
     <Reference Include="Microsoft.Extensions.WebEncoders"  />
     <Reference Include="Microsoft.AspNetCore.Components"  />
+    <Reference Include="Microsoft.AspNetCore.Components.Authorization"  />
   </ItemGroup>
 </Project>
